rename directory/ pkg
